+++ src/get_demographic_data.py
@@ -1,3 +1,4 @@
+import logging
 import os
 import pickle
 import shutil
@@ -7,29 +8,35 @@ import defopt
 import pandas as pd
 import tqdm
 import numpy as np
-from src.utils import Quarter, QuestionConfig, generate_quarters, read_demo_data, read_therapy_data
+from src.utils import (
+    Quarter,
+    QuestionConfig,
+    generate_quarters,
+    read_demo_data,
+    read_therapy_data,
+)
+
+logger = logging.getLogger("FAERS")
 
 
 def get_relevant_cases(fn_marked, config, nrows=None):
-    if fn_marked.endswith('csv'):
-        df_marked = pd.read_csv(
-            fn_marked,
-            nrows=nrows
-        )
+    if fn_marked.endswith("csv"):
+        df_marked = pd.read_csv(fn_marked, nrows=nrows)
     else:
-        df_marked = pickle.load(open(fn_marked, 'rb'))
+        df_marked = pickle.load(open(fn_marked, "rb"))
         if nrows is not None:
             df_marked = df_marked.head(nrows)
-    if 'caseid' not in df_marked.columns:
-        assert df_marked.index.name == 'caseid'
+    if "caseid" not in df_marked.columns:
+        assert df_marked.index.name == "caseid"
         df_marked.reset_index(inplace=True)
 
-    columns_bookkeeping = ['caseid']
-    columns_info = [c for c in df_marked if c.startswith('exposed ') or c.startswith('reacted ')]
+    columns_bookkeeping = ["caseid"]
+    columns_info = [
+        c for c in df_marked if c.startswith("exposed ") or c.startswith("reacted ")
+    ]
     dtypes = {c: str for c in columns_bookkeeping}
     for c in columns_info:
         dtypes[c] = bool
-
 
     df_marked = df_marked[columns_bookkeeping + columns_info]
     for c, dt in dtypes.items():
@@ -37,60 +44,55 @@ def get_relevant_cases(fn_marked, config, nrows=None):
 
     # columns_drugs = [f'drug {d}' for d in config.drugs]
     # columns_reactions = [f'reaction {r}' for r in config.reactions]
-    col_exposed = f'exposed {config.name}'
-    col_reacted = f'reacted {config.name}'
+    col_exposed = f"exposed {config.name}"
+    col_reacted = f"reacted {config.name}"
     drug_true = df_marked[col_exposed]
     reaction_true = df_marked[col_reacted]
     drug_naive = ~df_marked[col_exposed]
     ret = df_marked[columns_bookkeeping].copy()
-    ret['true_true'] = (drug_true & reaction_true).values
-    ret['true_false'] = (drug_true & (~reaction_true)).values
-    ret['drug_naive_true'] = (drug_naive & reaction_true).values
-    ret['drug_naive_false'] = (drug_naive & (~reaction_true)).values
+    ret["true_true"] = (drug_true & reaction_true).values
+    ret["true_false"] = (drug_true & (~reaction_true)).values
+    ret["drug_naive_true"] = (drug_naive & reaction_true).values
+    ret["drug_naive_false"] = (drug_naive & (~reaction_true)).values
     return ret.loc[
         ret.true_true | ret.true_false | ret.drug_naive_true | ret.drug_naive_false
-        ]
+    ]
 
 
 def process_a_config(q_start, q_end, dir_marked_data, dir_raw_data, dir_out, config):
     os.makedirs(dir_out, exist_ok=True)
     quarters = list(generate_quarters(q_start, q_end))
     row = np.random.randint(1, 10)
-    for q in tqdm.tqdm(
-        quarters, position=row, leave=False,
-        desc=config.name
-    ):
-        fn_marked = os.path.join(dir_marked_data, f'{q}.pkl')
-        fn_demo = os.path.join(dir_raw_data, f'demo{q}.csv.zip')
-        fn_therapy = os.path.join(dir_raw_data, f'ther{q}.csv.zip')
+    for q in tqdm.tqdm(quarters, position=row, leave=False, desc=config.name):
+
+        dir_out_curr = os.path.join(dir_out, config.name)
+        os.makedirs(dir_out_curr, exist_ok=True)
+        fn_out = os.path.join(dir_out_curr, f"{q}.csv.zip")
+        if os.path.exists(fn_out):
+            logger.debug(f"Skipping {q} because {fn_out} already exists")
+            continue
+        fn_marked = os.path.join(dir_marked_data, f"{q}.pkl")
+        fn_demo = os.path.join(dir_raw_data, f"demo{q}.csv.zip")
+        fn_therapy = os.path.join(dir_raw_data, f"ther{q}.csv.zip")
         df_cases = get_relevant_cases(fn_marked, config)
         df_demo = read_demo_data(fn_demo)
         df_therapy = read_therapy_data(fn_therapy)
-        df_cases = df_cases.merge(
-            df_demo,
-            on='caseid',
-            how='left'
-        ).merge(
-            df_therapy,
-            on='caseid',
-            how='left'
+        df_cases = df_cases.merge(df_demo, on="caseid", how="left").merge(
+            df_therapy, on="caseid", how="left"
         )
-        dir_out_curr = os.path.join(dir_out, config.name)
-        os.makedirs(dir_out_curr, exist_ok=True)
-        fn_out = os.path.join(dir_out_curr, f'{q}.csv.zip')
-        df_cases.to_csv(fn_out, index=False, compression='zip')
+        df_cases.to_csv(fn_out, index=False, compression="zip")
 
 
 def main(
-        *,
-        year_q_from,
-        year_q_to,
-        dir_marked_data,
-        dir_raw_data,
-        dir_config,
-        dir_out,
-        threads=4,
-        clean_on_failure=False
+    *,
+    year_q_from,
+    year_q_to,
+    dir_marked_data,
+    dir_raw_data,
+    dir_config,
+    dir_out,
+    threads=4,
+    clean_on_failure=False,
 ):
     """
     :param str year_q_from:
@@ -124,13 +126,16 @@ def main(
                 tqdm.tqdm(
                     pool.imap(
                         lambda config: process_a_config(
-                            q_start=q_from, q_end=q_to, dir_marked_data=dir_marked_data,
-                            dir_raw_data=dir_raw_data, dir_out=dir_out,
-                            config=config
+                            q_start=q_from,
+                            q_end=q_to,
+                            dir_marked_data=dir_marked_data,
+                            dir_raw_data=dir_raw_data,
+                            dir_out=dir_out,
+                            config=config,
                         ),
-                        configs
+                        configs,
                     ),
-                    total=len(configs)
+                    total=len(configs),
                 )
             )
     except Exception as err:
@@ -139,5 +144,5 @@ def main(
         raise err
 
 
-if __name__ == '__main__':
+if __name__ == "__main__":
     defopt.run(main)
